Added support for named POSIX semaphores (vislib::sys::Semaphore. It is recommended to use these instead of System V semaphores (vislib::sys::IPCSemaphore) for inter-process synchronisation.
git-svn-id: https://svn.vis.uni-stuttgart.de/utilities/vislib/trunk@482 3ec011a9-ccf5-4689-8f2f-5f4aa021d377

+++ test/testthread.cpp
@@ -303,7 +303,7 @@ void TestThread(void) {
 
 
     std::cout << std::endl << "Manual-reset event tests" << std::endl;
-    Semaphore resetSem(0);
+    Semaphore resetSem((long) 0);
     udSignal.evt = &evtManual;
     udSignal.resetSem = &resetSem;
     udWait.evt = &evtManual;
